add drupal converters

+++ wordpress/convert_articles_to_prismic_docs.js
@@ -2,6 +2,7 @@
 import fs from 'fs';
 import mkdirp from 'mkdirp';
 import {articleToPrismicParser} from './parsers';
+import keeps from './keeps';
 
 mkdirp.sync('./.dist/prismic');
 
@@ -12,11 +13,14 @@ fs.readdir('./.dist/articles', (err, files) => {
       if (err) throw err;
       const obj = JSON.parse(data);
       const slug = obj.url.replace('/articles/', '');
-      const prismicDoc = JSON.stringify(articleToPrismicParser(slug, obj, i));
-      fs.writeFile(`./.dist/prismic/${slug}.json`, prismicDoc, (err) => {
-        if (err) throw (err);
-        // console.log('The file was saved!');
-      });
+      const prismicDoc = JSON.stringify(articleToPrismicParser(slug, obj, i), null, 2);
+
+      if (keeps.includes(slug)) {
+        fs.writeFile(`./.dist/prismic/${slug}.json`, prismicDoc, (err) => {
+          if (err) throw (err);
+          // console.log('The file was saved!');
+        });
+      }
     });
   });
 });
